# msgidparse: make whitespace separating message ids optional