--- a/src/msgidparse.rs
+++ b/src/msgidparse.rs
@@ -53,14 +53,23 @@ impl fmt::Display for MessageIdList {
 /// ```
 pub fn msgidparse(ids: &str) -> Result<MessageIdList, MailParseError> {
     let mut msgids = Vec::new();
-    for id in ids.split_whitespace() {
-        if !id.starts_with('<') {
+
+    // The remaining section of the header, not yet chomped
+    let mut remaining = ids.trim_start();
+    // While we have some value of the header remaining
+    while remaining.len() > 0 {
+        // The next character should be the start of a Message ID
+        if !remaining.starts_with('<') {
             return Err(MailParseError::Generic("Message IDs must start with <"));
         }
-        if !id.ends_with('>') {
-            return Err(MailParseError::Generic("Message IDs must end with >"));
-        }
-        msgids.push(id[1..id.len() - 1].to_string());
+        // The ID ends at the next '>'
+        let end_index = remaining
+            .find('>')
+            .ok_or_else(|| MailParseError::Generic("Message IDs must end with >"))?;
+        msgids.push(remaining[1..end_index].to_string());
+
+        // Chomp the part of the string we just processed, and any trailing whitespace
+        remaining = &remaining[end_index + 1..].trim_start();
     }
     Ok(MessageIdList(msgids))
 }
@@ -102,6 +111,25 @@ mod tests {
                 "msg_one@foo.com".to_string(),
                 "msg_two@bar.com".to_string(),
                 "msg_three@qux.com".to_string(),
+            ])
+        );
+
+        // Non whitespace separator tests
+        assert_eq!(
+            msgidparse("<msg_one@foo.com><msg_two@bar.com>")
+                .expect("Multiple references, no whitespace"),
+            MessageIdList(vec![
+                "msg_one@foo.com".to_string(),
+                "msg_two@bar.com".to_string(),
+            ])
+        );
+        assert_eq!(
+            msgidparse("<msg_one@foo.com><msg_two@bar.com> <msg_three@spam.com> ")
+                .expect("Mixed whitespace/non-whitespace separator"),
+            MessageIdList(vec![
+                "msg_one@foo.com".to_string(),
+                "msg_two@bar.com".to_string(),
+                "msg_three@spam.com".to_string(),
             ])
         );
     }