Closes #72 

This PR:

- adds a benchmark using [criterion](https://github.com/bheisler/criterion.rs) for the current `msgidparse` function
- covers the existing `msgidparse` function with unit tests
- updates the implementation of `msgidparse` to make separator whitespace optional, as per  [RFC5322](https://tools.ietf.org/html/rfc5322#section-3.6.4) 

## Benchmark results

**No significant decrease** in speed was seen for short or long inputs to `msgidparse`.

### master

```log
msgidparse single id    time:   [74.461 ns 74.700 ns 74.958 ns]
msgidparse many ids     time:   [8.0999 us 8.1436 us 8.1853 us]
```

### This PR

```log
msgidparse single id    time:   [75.269 ns 75.839 ns 76.397 ns]
                        change: [+0.1514% +0.8757% +1.6172%] (p = 0.02 < 0.05)
                        Change within noise threshold.
msgidparse many ids     time:   [8.1968 us 8.2692 us 8.3627 us]
                        change: [-0.1450% +1.0312% +2.5808%] (p = 0.15 > 0.05)
                        No change in performance detected.
```