@tommilligan 2020-07-02
@staktrace updated with your comments addressed.